added wqork with migrations + fixed bug with favorites

+++ core/models/Favorites.py
@@ -1,10 +1,9 @@
-from sqlalchemy import ForeignKey, Table, Column, Integer
-from sqlalchemy.orm import mapped_column
+from sqlalchemy import ForeignKey, Table, Column, Integer, BigInteger
 from .Base import Base
 
 favorites_table = Table(
     "favorites",
     Base.metadata,
-    Column("user_id",    Integer, ForeignKey("users.id"),    primary_key=True),
+    Column("user_id",    BigInteger, ForeignKey("users.id"),    primary_key=True),
     Column("product_id", Integer, ForeignKey("products.id"), primary_key=True),
 )
